requirements.txt
flickrapi>=2.4.0
deepface>=0.0.98
requests>=2.32.0
python-dotenv>=1.0.0
